run-mirofish-with-regolo-a-complete-integration-guide/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
mem0ai>=0.1.0
